Added Comments

+++ Migrations/AmazonBooksDBContextModelSnapshot.cs
@@ -41,8 +41,8 @@ namespace BC_IS413_Assignment5.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("Price")
-                        .HasColumnType("int");
+                    b.Property<double>("Price")
+                        .HasColumnType("float");
 
                     b.Property<string>("Publisher")
                         .IsRequired()
